A little more general model generation
